extract platform related helper methods
these really aren’t related to configuration, so move them into their
own class to start to clean things up
