+++ features/support/path_helpers.rb
@@ -25,7 +25,7 @@ module PathHelpers
 
   def preseve_cmds_in_path(cmds, tmpbindir)
     cmds.each do |cmd|
-      whichcmd = Lolcommits::Configuration.command_which(cmd)
+      whichcmd = Lolcommits::Helper::Platform.command_which(cmd)
       unless whichcmd.nil?
         FileUtils.ln_s whichcmd, File.join(tmpbindir, File.basename(whichcmd))
       end
